requirements.txt
fastapi==0.103.1
uvicorn
# transformers==4.32.1
farm-haystack[faiss-gpu]
openpyxl==3.1.2